client, server: Ignore SIGPIPE when writing to a closed socket
Writing to a closed socket generates SIGPIPE, which by default silently
terminates the program. The previous solution was to supply the
`MSG_NOSIGNAL` flag to `send`, which disables SIGPIPE generation.

However, this solution is not portable, as macOS does not define
`MSG_NOSIGNAL`.

Instead, we use `sigaction` to ignore SIGPIPE in the whole process.